ToolBox: move color picker
